Feat: index, autoindex #20
